feat(demo): Load littlefs image file into simulated flash
